+++ docs/CHANGELOG.rst
@@ -1,6 +1,13 @@
 ChangeLog
 =========
 
+0.3.0 (2022-16-11)
+------------------
+* Pubtools CGW can support nested structure of data in YAML file
+* Changed the YAML field `state` to `action` to denote more appropriate operation
+* File `order` in YAML will be auto incremented by 10 for nested data structure
+* Updated documentation
+
 0.2.2 (2022-10-11)
 ------------------